requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.20
bcrypt==4.2.0
pyjwt==2.9.0
anthropic==0.97.0
reportlab==4.4.4
python-pptx==1.0.2
